Moved bump version task to specific package
