+++ src/main/groovy/org/shipkit/internal/gradle/version/VersioningPlugin.java
@@ -5,7 +5,7 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
-import org.shipkit.gradle.BumpVersionFileTask;
+import org.shipkit.gradle.version.BumpVersionFileTask;
 import org.shipkit.internal.gradle.git.GitPlugin;
 import org.shipkit.internal.gradle.util.StringUtil;
 import org.shipkit.internal.gradle.util.TaskMaker;
